Big commit, core logic for arduino.
This commit adds the main logic for the arduino int the 'main' folder.
Included is the main file, as well as a status-publisher, the logic for the ultrasonic-sensor and its publisher
as well as the logic for the step-motor that moves the door with its
subscriber.

Also included are smaller snippets, which where used for testing and are
now combined into the main logic. these can be found in exapmles.